Fixed division by zero in variant calling
When more than one alternative base are available, Strelka reports all
of them in "passed" vcf, even if one or more than one did not pass the
filters. This can cause a 0 value for both reference and alternative
coverages, leading to a total coverage equal to zero and an invalid
normalization. The missing filtering from Strelka is not straightforward
to fix, but at least we prevented the invalid division.