Implemented new unlink system
Now the Analysis handles the `can_unlink` flag. The Executor checks if
this flag is set before performing any unlinking operation. If
`can_unlink` is not set and the output files are overwritten, then it is
possible to unlink the following files and the flag is reset to `True`.
The Starter always set `can_unlink` to `False`, therefore the initial
files cannot be deleted.