Added Executor.override_last_operation_filename
This is an utility function to help change the value of the last
operation filename without worrying of making a mess. It works only when
one filename is available from the last operation, and it is very useful
when then file created was not 'barcodable' (otherwise an executor with
a lambda could be used).