Better support of non-mandatory software
This is just the first step, there are many different things that still
need to be handled. Nevertheless, at to date we only have a full
pipeline for mutation analysis, therefore all the software related to
the steps for this aim are mandatory. What should be not mandator is the
RNA-seq related software, like STAR.

 * `star_index` and `star_annotation` now belong to
   `Config.optional_files`
 * `Config.check_files` now calls `Config.check_file`, a simple function
   to check a single file
 * Added `Config.check_star_files`
 * `set_aligner_param` now also needs `files_checker`, an iterable of
   optional functions, needed, when available, to perform additional
   checks using the `Config`
 * Removed useless check of the aligner parameter
 * Mutation analysis must be skipped in case of RNA-seq samples inside
   runner
 * Renamed `Runner._run` to `Runner._run_mutation_analysis`