Fixed STAR behaviour
STAR was not using the current organism string for neither the index directory or the output filename. To fix this, additional parameters have been added to the config (`star_annotation` is now `features_{organism_str}` because it was not related to STAR but it is a normal genomic features file). Moreover, _STAR_ filename output cannot work with the barcoding, so there is an intermediate renaming phase which does not need the parsing of the filename and produces a valid one.

The `Executor` now exports the current `config` to the passed parameters, in order to perform some advanced replacement.

FIX: added `Mapping.create_bam_index` for the creation of `bai`s. This must be run after the rename phase in `Mapping.add_bam_groups`